move camera and window outside of window folder
